Posting: make the record-existence DAO a Spring @Repository
Drops the IReferencedRecordDAO/ISingletonService pair in favour of a single
@Repository next to its siblings in org.adempiere.ad.table; PostingService
resolves it via SpringContextHolder.lazyBean like its other Spring
collaborators.
