--- backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/table/api/IReferencedRecordDAO.java
@@ -1,9 +1,0 @@
-package org.adempiere.ad.table.api;
-
-import de.metas.util.ISingletonService;
-import org.adempiere.util.lang.impl.TableRecordReference;
-
-public interface IReferencedRecordDAO extends ISingletonService
-{
-	boolean exists(TableRecordReference recordRef);
-}
