+++ backend/de.metas.adempiere.adempiere/base/src/test/java/org/adempiere/ad/table/ReferencedRecordDAOTest.java
@@ -1,7 +1,5 @@
-package org.adempiere.ad.table.api.impl;
+package org.adempiere.ad.table;
 
-import de.metas.util.Services;
-import org.adempiere.ad.table.api.IReferencedRecordDAO;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.test.AdempiereTestHelper;
 import org.adempiere.util.lang.impl.TableRecordReference;
@@ -11,15 +9,37 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+/*
+ * #%L
+ * de.metas.adempiere.adempiere.base
+ * %%
+ * Copyright (C) 2026 metas GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 2 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program. If not, see
+ * <http://www.gnu.org/licenses/gpl-2.0.html>.
+ * #L%
+ */
+
 public class ReferencedRecordDAOTest
 {
-	private IReferencedRecordDAO referencedRecordDAO;
+	private ReferencedRecordDAO referencedRecordDAO;
 
 	@BeforeEach
 	void beforeEach()
 	{
 		AdempiereTestHelper.get().init();
-		referencedRecordDAO = Services.get(IReferencedRecordDAO.class);
+		referencedRecordDAO = ReferencedRecordDAO.newInstanceForUnitTesting();
 	}
 
 	@Test
